Get data from a specific location from the Hystreet Project via Hystreet API

Get data from a specific location from the Hystreet Project via Hystreet API

get_hystreet_station_data(hystreetId = NULL, query = NULL, API_token = NULL)

Arguments

hystreetId
integer (required): ID of the requested station. See get_hystreet_locations() for an overview of available IDs.

query
list (optional): A list with queries. Up do date the following queries are supported:

from: datetime of earliest measurement (default: today 00:00:00:): e.g. "2018-10-01 12:00:00" or "2018-10-01"

to : datetime of latest measurement (default: today 23:59:59): e.g. "2018-01-12 12:00:00" or "2018-12-01"

resolution: Resolution for the measurement grouping (default: hour): "day", "hour", "month", "week"

API_token
character (optional): API key to get access to Hystreet API

Value

data.frame with parsed data from hystreet API

Function version

0.0.2

Examples

# \dontrun{
## request data of the current day of station with hystreetId 71
get_hystreet_station_data(71)

#> $id
#> [1] 71
#> 
#> $name
#> [1] "Schusterstraße"
#> 
#> $city
#> [1] "Mainz"
#> 
#> $statistics
#> $statistics$today_count
#> [1] 6111
#> 
#> $statistics$timerange_count
#> [1] 7365
#> 
#> 
#> $metadata
#> $metadata$measured_from
#> [1] "2020-06-12T00:00:00.000+02:00"
#> 
#> $metadata$measured_to
#> [1] "2020-06-12T23:59:59.999+02:00"
#> 
#> $metadata$earliest_measurement_at
#> [1] "2018-05-01 02:00:00 CEST"
#> 
#> $metadata$latest_measurement_at
#> [1] "2020-06-12 15:14:57 CEST"
#> 
#> $metadata$ltr_label
#> NULL
#> 
#> $metadata$rtl_label
#> NULL
#> 
#> 
#> $measurements
#>              timestamp weather_condition temperature min_temperature
#> 1  2020-06-12 00:00:00       clear-night       15.36              NA
#> 2  2020-06-12 01:00:00       clear-night       14.83              NA
#> 3  2020-06-12 02:00:00       clear-night       14.52              NA
#> 4  2020-06-12 03:00:00       clear-night       14.07              NA
#> 5  2020-06-12 04:00:00       clear-night       13.36              NA
#> 6  2020-06-12 05:00:00       clear-night       13.25              NA
#> 7  2020-06-12 06:00:00         clear-day       13.97              NA
#> 8  2020-06-12 07:00:00         clear-day       15.07              NA
#> 9  2020-06-12 08:00:00         clear-day       16.21              NA
#> 10 2020-06-12 09:00:00         clear-day       17.81              NA
#> 11 2020-06-12 10:00:00         clear-day       19.84              NA
#> 12 2020-06-12 11:00:00         clear-day       21.45              NA
#> 13 2020-06-12 12:00:00         clear-day       22.83              NA
#> 14 2020-06-12 13:00:00         clear-day       24.18              NA
#> 15 2020-06-12 14:00:00 partly-cloudy-day       25.25              NA
#> 16 2020-06-12 15:00:00 partly-cloudy-day       26.39              NA
#> 17 2020-06-12 16:00:00         clear-day       27.21              NA
#> 18 2020-06-12 17:00:00         clear-day       27.81              NA
#> 19 2020-06-12 18:00:00         clear-day       27.91              NA
#> 20 2020-06-12 19:00:00         clear-day       27.29              NA
#> 21 2020-06-12 20:00:00         clear-day       26.16              NA
#> 22 2020-06-12 21:00:00         clear-day       24.99              NA
#> 23 2020-06-12 22:00:00       clear-night       23.43              NA
#> 24 2020-06-12 23:00:00       clear-night       22.13              NA
#>    pedestrians_count unverified
#> 1                 11       TRUE
#> 2                  4       TRUE
#> 3                 13       TRUE
#> 4                  2       TRUE
#> 5                 10       TRUE
#> 6                 10       TRUE
#> 7                 26       TRUE
#> 8                 64       TRUE
#> 9                213       TRUE
#> 10               558       TRUE
#> 11               987       TRUE
#> 12              1330       TRUE
#> 13              1501       TRUE
#> 14              1282       TRUE
#> 15              1068       TRUE
#> 16               286       TRUE
#> 17                 0       TRUE
#> 18                 0       TRUE
#> 19                 0       TRUE
#> 20                 0       TRUE
#> 21                 0       TRUE
#> 22                 0       TRUE
#> 23                 0       TRUE
#> 24                 0       TRUE
#> 
#> $incidents
#> list()
#> 
#> $permissions
#> list()
#> 

## request data of December 2018 with resolution "day"
get_hystreet_station_data(
   hystreetId = 71,
   query = list(from = "2018-12-01", to = "2018-12-31", resolution = "day"))

#> $id
#> [1] 71
#> 
#> $name
#> [1] "Schusterstraße"
#> 
#> $city
#> [1] "Mainz"
#> 
#> $statistics
#> $statistics$today_count
#> [1] 6111
#> 
#> $statistics$timerange_count
#> [1] 484991
#> 
#> 
#> $metadata
#> $metadata$measured_from
#> [1] "2018-12-01"
#> 
#> $metadata$measured_to
#> [1] "2018-12-31"
#> 
#> $metadata$earliest_measurement_at
#> [1] "2018-05-01 02:00:00 CEST"
#> 
#> $metadata$latest_measurement_at
#> [1] "2020-06-12 15:14:57 CEST"
#> 
#> $metadata$ltr_label
#> NULL
#> 
#> $metadata$rtl_label
#> NULL
#> 
#> 
#> $measurements
#>     timestamp weather_condition temperature min_temperature pedestrians_count
#> 1  2018-12-01              <NA>          NA              NA             31451
#> 2  2018-12-02              rain       12.58            7.70              9422
#> 3  2018-12-03              rain       13.36           10.05             10161
#> 4  2018-12-04              rain       10.56            1.00             17426
#> 5  2018-12-05 partly-cloudy-day        5.97            1.04             17624
#> 6  2018-12-06              rain       10.69            4.98             17698
#> 7  2018-12-07              rain       12.25            7.47             18157
#> 8  2018-12-08              rain        9.29            6.56             27396
#> 9  2018-12-09              rain        9.68            6.55              9035
#> 10 2018-12-10              rain        6.51            3.94             14335
#> 11 2018-12-11 partly-cloudy-day        6.45            3.03             17087
#> 12 2018-12-12 partly-cloudy-day        5.47            2.97             18475
#> 13 2018-12-13 partly-cloudy-day        3.35            0.25             18412
#> 14 2018-12-14            cloudy        2.56            0.30             20732
#> 15 2018-12-15            cloudy        1.24           -1.58             31730
#> 16 2018-12-16 partly-cloudy-day        0.82           -2.46              9679
#> 17 2018-12-17 partly-cloudy-day        4.03            0.36             17048
#> 18 2018-12-18 partly-cloudy-day        6.90            1.58             21120
#> 19 2018-12-19 partly-cloudy-day        5.47            1.33             20423
#> 20 2018-12-20              rain        7.91            3.85             22773
#> 21 2018-12-21              rain       12.45            6.57             21628
#> 22 2018-12-22              rain       13.31            8.32             29328
#> 23 2018-12-23              rain       11.28            6.64              7307
#> 24 2018-12-24              rain       12.93           -0.04              6545
#> 25 2018-12-25 partly-cloudy-day        2.91           -1.51              2801
#> 26 2018-12-26 partly-cloudy-day       -0.85           -3.83              2835
#> 27 2018-12-27 partly-cloudy-day       -0.15           -2.02             11600
#> 28 2018-12-28            cloudy        0.41           -1.07             12276
#> 29 2018-12-29            cloudy        4.88           -0.21             15924
#> 30 2018-12-30              rain        8.36            4.95              4533
#> 31 2018-12-31              rain        9.56            5.62                30
#>    unverified
#> 1       FALSE
#> 2       FALSE
#> 3       FALSE
#> 4       FALSE
#> 5       FALSE
#> 6       FALSE
#> 7       FALSE
#> 8       FALSE
#> 9       FALSE
#> 10      FALSE
#> 11      FALSE
#> 12      FALSE
#> 13      FALSE
#> 14      FALSE
#> 15      FALSE
#> 16      FALSE
#> 17      FALSE
#> 18      FALSE
#> 19      FALSE
#> 20      FALSE
#> 21      FALSE
#> 22      FALSE
#> 23      FALSE
#> 24      FALSE
#> 25      FALSE
#> 26      FALSE
#> 27      FALSE
#> 28      FALSE
#> 29      FALSE
#> 30      FALSE
#> 31      FALSE
#> 
#> $incidents
#> list()
#> 
#> $permissions
#> list()
#> 

 # }