Get statistics from the Hystreet Project via Hystreet API

Get statistics from the Hystreet Project via Hystreet API

get_hystreet_stats(verbose = FALSE, API_token = NULL)

Arguments

verbose
logical (optional): Show more verbose output?

API_token
character (optional): API key to get access to Hystreet API

Value

data.frame with parsed data from hystreet API

Function version

0.0.2

Examples

# \dontrun{
 get_hystreet_stats(TRUE)

#>      id                             name                city today_count
#> 1   246              Westenhellweg (Ost)            Dortmund       14613
#> 2    59                     Goethestraße      Frankfurt a.M.        1921
#> 3    57                  Schlösserstraße              Erfurt        7941
#> 4    68                     Petersstraße             Leipzig       12715
#> 5   256                Kirchgasse (Nord)           Wiesbaden       12681
#> 6    86               Löhrstraße (Mitte)             Koblenz       12561
#> 7   107                Krahnstraße (Süd)           Osnabrück        6013
#> 8   206         Kröpeliner Straße (West)             Rostock       11930
#> 9    77                Königstraße (Süd)           Stuttgart       18490
#> 10  150                     Zeil (Mitte)      Frankfurt a.M.       26979
#> 11   53             Schadowstraße (West)          Düsseldorf       15454
#> 12  254                    Zöllnerstraße               Celle        3826
#> 13   47             Schildergasse (West)                Köln       25814
#> 14  125                 Kettwiger Straße               Essen       13319
#> 15  167                  Zeppelinstraße                 Köln       13006
#> 16  138                       Brotstraße               Trier       10337
#> 17   73                 Neuhauser Straße             München       27549
#> 18  166         Maximilianstraße (Mitte)             München        3695
#> 19   72                  Stadthausstraße               Mainz        7993
#> 20  257            Holstenstraße (Mitte)                Kiel        7916
#> 21  104              Ernst-Ludwig-Straße           Darmstadt        7296
#> 22  155              Georgstraße  (West)           Osnabrück        4925
#> 23  129                      Schuhstraße        Braunschweig        4727
#> 24  117                    Ludgeristraße             Münster       12860
#> 25  126                      Neue Straße        Braunschweig        3173
#> 26   71                   Schusterstraße               Mainz        6111
#> 27  132                       Langer Hof        Braunschweig        4597
#> 28  154                       Poststraße           Wuppertal       14539
#> 29  131                             Sack        Braunschweig        4612
#> 30  130                             Damm        Braunschweig        4708
#> 31  128                       Hutfiltern        Braunschweig        5160
#> 32  127                       Poststraße        Braunschweig        5527
#> 33  151                      Stiftstraße           Stuttgart        5321
#> 34   87                       Poststraße                Bonn       10184
#> 35   58        Große Bockenheimer Straße      Frankfurt a.M.        8395
#> 36  157                     Mittelstraße          Düsseldorf        7440
#> 37   91                       Neuer Wall             Hamburg        7240
#> 38  253                        Neustraße             Bocholt        6860
#> 39  260          Schadowstraße (Südwest)          Düsseldorf       10986
#> 40  143     Kaiser-Joseph-Straße (Mitte)            Freiburg       14633
#> 41   56       Königsallee Ostseite (Süd)          Düsseldorf        7689
#> 42   56       Königsallee Ostseite (Süd)          Düsseldorf        7689
#> 43  139                   Fleiner Straße           Heilbronn       14251
#> 44  245         Hindenburgstraße (Mitte)     Mönchengladbach       11125
#> 45  141               Hauptstraße (West)          Heidelberg       12685
#> 46  170            Adalbertstraße (West)              Aachen       12994
#> 47   52                      Sternstraße                Bonn       12678
#> 48  210                    Achternstraße           Oldenburg        6767
#> 49   78                    Münsterplatz                  Ulm        8921
#> 50  249            Bahnhofstraße (Mitte)              Brilon        2038
#> 51  251                       Annastraße            Augsburg        7902
#> 52  248             Bahnhofstraße (West)              Brilon        2352
#> 53  100                       Rothenburg             Münster       10086
#> 54  147                     Ludwigstraße              Passau       11162
#> 55   79               Kirchgasse (Mitte)           Wiesbaden       16740
#> 56  161                 Stresemannstraße           Bielefeld       11446
#> 57   96                    Planken (Ost)            Mannheim        7635
#> 58   88                     Breite Gasse            Nürnberg       12085
#> 59  116        Herrenteichsstraße (West)           Osnabrück        3847
#> 60  108             Große Straße (Mitte)           Osnabrück       10897
#> 61   65                  Karmarschstraße            Hannover        9331
#> 62   51                   Remigiusstraße                Bonn       13375
#> 63   64              Große Packhofstraße            Hannover        8721
#> 64  171                      Hauptstraße   Bergisch-Gladbach        8360
#> 65  258             Mittelstraße (Mitte)               Lemgo        5012
#> 66  114         Herrenteichsstraße (Ost)           Osnabrück        2155
#> 67  261           Flinger Straße (Mitte)          Düsseldorf       11911
#> 68  111                       Jürgensort           Osnabrück        5077
#> 69   75                    Bahnhofstraße         Saarbrücken       12715
#> 70   55             Flinger Straße (Ost)          Düsseldorf       12884
#> 71  259            Schadowstraße (Mitte)          Düsseldorf        6408
#> 72   85                         Am Brand               Mainz       10998
#> 73  208                        Hoher Weg          Hildesheim           0
#> 74  244             Weender Straße (Süd)           Göttingen        6690
#> 75   98 Tauentzienstraße Südseite (West)              Berlin           0
#> 76   69                       Hainstraße             Leipzig       13251
#> 77  207            Breite Straße (Mitte)              Lübeck        9972
#> 78  135                  Schönbornstraße            Würzburg       16830
#> 79  162            Bahnhofstraße (Mitte)           Bielefeld       11527
#> 80  164             Prager Straße (Nord)             Dresden       13029
#> 81  148             Werner-Senger-Straße             Limburg        9119
#> 82   63                      Georgstraße            Hannover       19364
#> 83  153               Große Straße (Süd)           Osnabrück       10242
#> 84  146                     Hirschstraße                 Ulm       16412
#> 85  134                    Wilhelmstraße          Reutlingen       11335
#> 86   90                  Karolinenstraße            Nürnberg       15535
#> 87  160              Schildergasse (Ost)                Köln       12421
#> 88  102     Bürgermeister-Fischer-Straße            Augsburg        9502
#> 89  205              Ehrenstraße (Mitte)                Köln        9925
#> 90   54              Schadowstraße (Ost)          Düsseldorf       12186
#> 91   95                  Planken (Mitte)            Mannheim       14330
#> 92   81                   Spitalerstraße             Hamburg       14407
#> 93   81                   Spitalerstraße             Hamburg       14407
#> 94   76              Königstraße (Mitte)           Stuttgart       21395
#> 95   93                   Große Bleichen             Hamburg        6021
#> 96   93                   Große Bleichen             Hamburg        6021
#> 97   84                     Rathausgasse            Freiburg        5887
#> 98   97    Kurfürstendamm Südseite (Ost)              Berlin        7545
#> 99   74                   Schützenstraße             München        3813
#> 100 165                  Kaufingerstraße             München       29299
#> 101  49            Westenhellweg (Mitte)            Dortmund       16624
#> 102 168               Hohe Straße (Nord)                Köln       17008
#> 103 158                Hohe Straße (Süd)                Köln       18680
#> 104 109    Krahnstraße (Mitte, Altstadt)           Osnabrück        5764
#> 105 123                     Ludwigstraße          Ingolstadt        8306
#> 106 103                  Schuchardstraße           Darmstadt        6871
#> 107 209                     Lange Straße           Oldenburg        6394
#> 108  94   Kurfürstendamm Nordseite (Ost)              Berlin        4181
#> 109  94   Kurfürstendamm Nordseite (Ost)              Berlin        4181
#> 110 122                       Seltersweg              Gießen        9614
#> 111 252                             Holm           Flensburg       10587
#> 112 140                    Fleischstraße               Trier        5206
#> 113 110            Johannisstraße (Nord)           Osnabrück        4328
#> 114  50                Ehrenstraße (Ost)                Köln        8594
#> 115  89                     Kaiserstraße            Nürnberg        5927
#> 116 159            Schildergasse (Mitte)                Köln       26710
#> 117 163              Bahnhofstraße (Süd)           Bielefeld       16702
#> 118 145               Rheinstraße (West)             Krefeld         164
#> 119 144               Hochstraße (Mitte)             Krefeld       10091
#> 120  48              Hohe Straße (Mitte)                Köln       18768
#> 121 118                     Grüner Markt             Bamberg       15918
#> 122  92                    Jungfernstieg             Hamburg       11487
#> 123 142                Hauptstraße (Süd)            Erlangen        5823
#> 124 250                 Bürgerturmstraße Biberach an der Riß        1996
#> 125 169                    Westernstraße           Paderborn       13520
#> 126 172                      Obernstraße              Bremen        7533
#> 127  67             Kaiserstraße (Mitte)           Karlsruhe       16982

# }